Fix bug on updating new item, and make delete button an option

+++ KCMCPharmacy.php
@@ -39,6 +39,8 @@ if (isset($_GET['Delete'])) {
 }
 
 if (isset($_POST['ChangeItem'])) {
+	$Patient[0]=$_POST['PatientNo'];
+	$Patient[1]=$_POST['BranchNo'];
 	$sql="SELECT price
 				FROM prices
 				WHERE stockid='".$_POST['StockID']."'
@@ -578,7 +580,11 @@ if (isset($Patient)) {
 			echo '<td>&nbsp;' . _('Quantity') . ' - ';
 			echo '&nbsp;' . $_SESSION['Items'][$i]['Quantity'];
 			echo '&nbsp;@&nbsp;'.number_format($_SESSION['Items'][$i]['Price'],0).' '.$_SESSION['CompanyRecord']['currencydefault'].'</td>';
-			echo '<td><a href="' . $_SERVER['PHP_SELF'] . '?Delete=' . $i . '&Patient='.$Patient[0].'&Branch='.$Patient[1].'">' . _('Delete') . '</a></td></tr>';
+			if ($_SESSION['CanViewPrices']==1){
+				echo '<td><a href="' . $_SERVER['PHP_SELF'] . '?Delete=' . $i . '&Patient='.$Patient[0].'&Branch='.$Patient[1].'">' . _('Delete') . '</a></td></tr>';
+			} else {
+				echo '<td>' . _('Delete') . '</td></tr>';
+			}
 			DB_data_seek($result,0);
 			echo '<tr><td>';
 		}
