+++ KCMCRadiology.php
@@ -39,6 +39,8 @@ if (isset($_GET['Delete'])) {
 }
 
 if (isset($_POST['ChangeItem'])) {
+	$Patient[0]=$_POST['PatientNo'];
+	$Patient[1]=$_POST['BranchNo'];
 	$sql="SELECT price
 				FROM prices
 				WHERE stockid='".$_POST['StockID']."'
@@ -549,7 +551,11 @@ if (isset($Patient)) {
 				}
 			}
 			echo '<td>&nbsp;@&nbsp;'.locale_money_format($_SESSION['Items'][$i]['Price'],$_SESSION['CompanyRecord']['currencydefault']).' '.$_SESSION['CompanyRecord']['currencydefault'].'</td>';
-			echo '<td><a href="' . $_SERVER['PHP_SELF'] . '?Delete=' . $i . '&Patient='.$Patient[0].'&Branch='.$Patient[1].'">' . _('Delete') . '</a></td></tr>';
+			if ($_SESSION['CanViewPrices']==1){
+				echo '<td><a href="' . $_SERVER['PHP_SELF'] . '?Delete=' . $i . '&Patient='.$Patient[0].'&Branch='.$Patient[1].'">' . _('Delete') . '</a></td></tr>';
+			} else {
+				echo '<td>' . _('Delete') . '</td></tr>';
+			}
 			DB_data_seek($result,0);
 			echo '<tr><td>';
 		}
